move api_test and clean up unit tests
